backend/requirements.txt
flask
flask-cors
requests
openai
google-generativeai
